Lord-Of-The-Flies-Project
I have created my first ever website!!! I did this by teaching myself HTML code, CSS code, jekyll code words, and how to upload a website. In this website, there is a summary of the novel Lord of the flies, there are character descriptions for the characters. After that there is an image of a fake video game case of the pretend game, Lord Of The Flies. This was my first ever website for a school website!!!